add: change double on float for mTargetFps

+++ src/application.h
@@ -8,6 +8,6 @@ public:
 private:
 	sf::RenderWindow& mWindow = Window::instance();
 	sf::Clock mClock; 
-	double mTargetFps = 60.0;
+	float mTargetFps = 60.0;
 	sf::Time mTargetTps = sf::seconds((1.f / mTargetFps));
 };
